inlined content type add overloads and separated out partypeinfo to a new file.
