+++ src/DocumentFormat.OpenXml/Packaging/ISupportedRelationship.cs
@@ -4,16 +4,15 @@
 using DocumentFormat.OpenXml.Features;
 using System;
 
-namespace DocumentFormat.OpenXml.Packaging
-{
-    /// <summary>
-    /// Defines the interface for tagging a part that can add extensible parts.
-    /// </summary>
-    /// <typeparam name="TPart">Extensible part type that is supported by the implementing class.</typeparam>
+namespace DocumentFormat.OpenXml.Packaging;
+
+/// <summary>
+/// Defines the interface for tagging a part that can add extensible parts.
+/// </summary>
+/// <typeparam name="TPart">Extensible part type that is supported by the implementing class.</typeparam>
 #pragma warning disable CA1040 // Avoid empty interfaces
-    public interface ISupportedRelationship<TPart>
+public interface ISupportedRelationship<TPart>
 #pragma warning restore CA1040 // Avoid empty interfaces
-        where TPart : OpenXmlPart
-    {
-    }
+    where TPart : OpenXmlPart
+{
 }
